Move vendor tests to __tests__/
